requirements.txt
Django==2.1.2
python-stdnum==1.10
